eliminate need for -hostrpc from compiles
cherry-picked from amd-stg-open

splits libomptarget/hostrpc/src/hostrpc.cpp into a device and host
build.

This imbeds the host functions from libhostrpc.a into libomptarget.so
