+++ clang/lib/Driver/ToolChains/CommonArgs.cpp
@@ -658,7 +658,6 @@ bool tools::addOpenMPRuntime(ArgStringList &CmdArgs, const ToolChain &TC,
       CmdArgs.push_back("-lrt");
 
   if (IsOffloadingHost) {
-    CmdArgs.push_back("-lhostrpc");
     CmdArgs.push_back("-lomptarget");
   }
