+++ openmp/libomptarget/src/CMakeLists.txt
@@ -16,6 +16,7 @@ set(src_files
   api.cpp
   device.cpp
   interface.cpp
+  hostrpc.cpp
   rtl.cpp
   omptarget.cpp
 )
